refactoring my work to fit into the init refactor
