+++ rover-connectors-starter-main/.github/CODEOWNERS
@@ -1,0 +1,1 @@
+* @apollographql/team-growth
