+++ rover-connectors-starter-main/.vscode/extensions.json
@@ -1,0 +1,3 @@
+{
+    "recommendations": ["apollographql.vscode-apollo","redhat.vscode-yaml"]
+}
