+++ rover-connectors-starter-main/README.md
@@ -1,0 +1,14 @@
+## Introduction
+
+👋 Welcome\!
+
+`rover init` is designed to help you set up a new, federated GraphQL API using Apollo’s Rover CLI, [Apollo Connectors for REST](https://www.apollographql.com/docs/graphos/schema-design/connectors), and your existing REST APIs.
+
+This command:
+
+* Creates a set of template files in your current directory
+* Generates the credentials your GraphQL API will need to interact with Apollo’s GraphOS platform
+* Provides a starter schema to help you design your GraphQL API
+* Walks you through spinning up a local environment to test your API
+
+ℹ️ New to \`rover init\` and this repo? Jump to the [Next Steps](?tab=t.0#heading=h.ajx7mu1cuq8h) section to install and configure Rover. Otherwise, head over to \`getting-started.md\` for next steps.
