+++ src/command/init/mod.rs
@@ -4,17 +4,14 @@ mod states;
 mod template_operations;
 mod transitions;
 
+use std::path::PathBuf;
+
 use crate::options::{
     GraphIdOpt, ProjectNameOpt, ProjectOrganizationOpt, ProjectTypeOpt, ProjectUseCaseOpt,
 };
-mod project;
-
-use crate::options::{ProjectUseCase, ProjectUseCaseOpt};
 use crate::{RoverOutput, RoverResult};
 use clap::Parser;
 use rover_http::ReqwestService;
-use dialoguer::Input;
-use project::Project;
 use serde::Serialize;
 
 #[cfg(test)]
@@ -39,7 +36,7 @@ pub struct Init {
     graph_id: GraphIdOpt,
 
     #[clap(long, hide(true))]
-    path: Option<Utf8PathBuf>,
+    path: Option<PathBuf>,
 }
 
 impl Init {
@@ -48,7 +45,7 @@ impl Init {
         let http_service = ReqwestService::new(None, None)?;
 
         let creation_confirmed_option = Welcome::new()
-            .select_project_type(&self.project_type)?
+            .select_project_type(&self.project_type, &self.path)?
             .select_organization(&self.organization)?
             .select_use_case(&self.project_use_case)?
             .enter_project_name(&self.project_name)?
@@ -63,33 +60,7 @@ impl Init {
             }
             None => Ok(RoverOutput::EmptySuccess),
         }
-        let use_case = self.use_case_options.get_or_prompt_use_case()?;
-
-        match use_case {
-            ProjectUseCase::GraphQLTemplate => println!("\nComing soon!\n"),
-            ProjectUseCase::Connectors => {
-                let _project_name = Project::prompt_for_valid_project_name()?;
-                println!("\nComing soon!\n")
-            }
-        }
-
-        Ok(RoverOutput::EmptySuccess)
     }
 }
 
 pub use states::Welcome;
-fn is_valid_string(input: &str, max_length: usize, allowed_chars: &str) -> bool {
-    // Check length
-    if input.len() > max_length {
-        return false;
-    }
-
-    // Check characters
-    for char in input.chars() {
-        if !allowed_chars.contains(char) {
-            return false;
-        }
-    }
-
-    true
-}
