+++ src/options/project_name.rs
@@ -1,7 +1,10 @@
+use crate::RoverError;
 use crate::RoverResult;
 use clap::arg;
 use clap::Parser;
+use dialoguer::Input;
 use serde::{Deserialize, Serialize};
+use anyhow::anyhow;
 
 #[derive(Debug, Clone, Serialize, Deserialize, Parser, Default)]
 pub struct ProjectNameOpt {
@@ -9,24 +12,37 @@ pub struct ProjectNameOpt {
     pub project_name: Option<String>,
 }
 
-// TODO: Include Taylor's logic for generating a project name
 impl ProjectNameOpt {
     pub fn get_project_name(&self) -> Option<String> {
         self.project_name.clone()
     }
 
     pub fn prompt_project_name(&self) -> RoverResult<String> {
-        // TODO: Include Taylor's prompt, validation, and return project name logic here
-        let _prompt = "Name your GraphQL API";
         let default = self.suggest_default_name();
+        let max_length = 64;
+        let min_length = 2;
+        let allowed_chars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890-!@#$%^&*()_+<>/?\\[]{};: ";
 
-        // TODO: Include Taylor's logic for generating a project name
+        loop {
+            let input: Input<String> = Input::new().with_prompt("? Name your GraphQL API").default(default.to_string());
+            let name = input.interact_text()?.to_string();
 
-        Ok(default)
+            let valid_name =
+                self.validate_name(&name, allowed_chars, max_length, min_length);
+
+            match valid_name {
+                Ok(valid_name) => {
+                    return Ok(valid_name);
+                }
+                Err(invalid_name) => {
+                    println!("{invalid_name}")
+                }
+            }
+        }
     }
 
     fn suggest_default_name(&self) -> String {
-        "my-graphql-api".to_string()
+        "My API".to_string()
     }
 
     pub fn get_or_prompt_project_name(&self) -> RoverResult<String> {
@@ -37,12 +53,201 @@ impl ProjectNameOpt {
 
         self.prompt_project_name()
     }
+
+    pub fn validate_name(
+        &self,
+        input: &str,
+        allowed_chars: &str,
+        max_length: usize,
+        min_length: usize,
+    ) -> RoverResult<String> {
+        // Check length
+        if !validate_max_length(input, max_length) {
+            return Err(RoverError::new(anyhow!(
+                "Names must be a maximum of {} characters long.",
+                max_length
+            )));
+        }
+
+        if !validate_min_length(input, min_length) {
+            return Err(RoverError::new(anyhow!(
+                "Names must be a minimum of {} characters long.",
+                min_length
+            )));
+        }
+
+        // Check characters
+        if !validate_allowed_chars(input, allowed_chars) {
+            return Err(RoverError::new(anyhow!(
+                "Names may only contain the following characters: {}",
+                allowed_chars
+            )));
+        }
+
+        Ok(input.to_string())
+    }
 }
 
-// TODO: Add tests for interactive prompts and sad paths
+fn validate_max_length(input: &str, max_length: usize) -> bool {
+    if input.len() > max_length {
+        return false;
+    }
+
+    return true;
+}
+
+fn validate_min_length(input: &str, min_length: usize) -> bool {
+    if input.len() < min_length {
+        return false;
+    }
+
+    return true;
+}
+
+fn validate_allowed_chars(input: &str, allowed_chars: &str) -> bool {
+    for char in input.chars() {
+        if !allowed_chars.contains(char) {
+            return false;
+        }
+    }
+
+    return true;
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn test_validate_max_length_returns_false_when_input_length_is_greater_than_max_length() {
+        let input = "test";
+        let max_length = 3;
+        let result = validate_max_length(input, max_length);
+
+        assert_eq!(result, false);
+    }
+
+    #[test]
+    fn test_validate_max_length_returns_true_when_input_length_is_equal_to_max_length() {
+        let input = "test";
+        let max_length = 4;
+        let result = validate_max_length(input, max_length);
+
+        assert_eq!(result, true);
+    }
+
+    #[test]
+    fn test_validate_max_length_returns_true_when_input_length_is_less_than_max_length() {
+        let input = "test";
+        let max_length = 5;
+        let result = validate_max_length(input, max_length);
+
+        assert_eq!(result, true);
+    }
+
+    #[test]
+    fn test_validate_min_length_returns_false_when_input_length_is_less_than_min_length() {
+        let input = "t";
+        let min_length = 2;
+        let result = validate_min_length(input, min_length);
+
+        assert_eq!(result, false);
+    }
+
+    #[test]
+    fn test_validate_min_length_returns_true_when_input_length_is_equal_to_min_length() {
+        let input = "test";
+        let min_length = 4;
+        let result = validate_min_length(input, min_length);
+
+        assert_eq!(result, true);
+    }
+
+    #[test]
+    fn test_validate_min_length_returns_true_when_input_length_is_greater_than_min_length() {
+        let input = "test";
+        let min_length = 3;
+        let result = validate_min_length(input, min_length);
+
+        assert_eq!(result, true);
+    }
+
+    #[test]
+    fn test_validate_allowed_chars_returns_false_when_input_contains_invalid_characters() {
+        let input = "test";
+        let allowed_chars = "abc";
+        let result = validate_allowed_chars(input, allowed_chars);
+
+        assert_eq!(result, false);
+    }
+
+    #[test]
+    fn test_validate_allowed_chars_returns_true_when_input_contains_allowed_characters() {
+        let input = "test";
+        let allowed_chars = "tes";
+        let result = validate_allowed_chars(input, allowed_chars);
+
+        assert_eq!(result, true);
+    }
+
+    #[test]
+    fn test_validate_allowed_chars_returns_true_when_input_contains_allowed_special_characters() {
+        let input = "!@#$%^&*()";
+        let allowed_chars = "0123456789!@#$%^&*()";
+        let result = validate_allowed_chars(input, allowed_chars);
+
+        assert_eq!(result, true);
+    }
+
+    #[test]
+    fn test_validate_returns_error_when_input_has_invalid_length() {
+        let instance = ProjectNameOpt {
+            project_name: Some("my-project".to_string()),
+        };
+
+        let input = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890-!@#$%^&*()_+";
+        let allowed_chars =
+            "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890-!@#$%^&*()_+";
+        let max_length = 1;
+        let min_length = 0;
+
+        let result = instance.validate_name(input, allowed_chars, max_length, min_length);
+
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_validate_returns_error_when_input_has_invalid_chars() {
+        let instance = ProjectNameOpt {
+            project_name: Some("my-project".to_string()),
+        };
+
+        let input = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890-!@#$%^&*()_+";
+        let allowed_chars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890";
+        let max_length = 100;
+        let min_length = 1;
+
+        let result = instance.validate_name(input, allowed_chars, max_length, min_length);
+
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_validate_returns_ok_when_input_has_valid_chars_and_length() {
+        let instance = ProjectNameOpt {
+            project_name: Some("my-project".to_string()),
+        };
+
+        let input = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890-!@#$%^&*()_+";
+        let allowed_chars =
+            "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890-!@#$%^&*()_+";
+        let max_length = 100;
+        let min_length = 1;
+
+        let result = instance.validate_name(input, allowed_chars, max_length, min_length);
+
+        assert!(result.is_ok());
+    }
 
     #[test]
     fn test_get_project_name_with_preset_value() {
@@ -59,7 +264,7 @@ mod tests {
         let instance = ProjectNameOpt { project_name: None };
         let default_name = instance.suggest_default_name();
 
-        assert_eq!(default_name, "my-graphql-api");
+        assert_eq!(default_name, "My API");
     }
 
     #[test]
@@ -68,7 +273,7 @@ mod tests {
         let result = instance.prompt_project_name();
 
         assert!(result.is_ok());
-        assert_eq!(result.unwrap(), "my-graphql-api");
+        assert_eq!(result.unwrap(), "My API");
     }
 
     #[test]
@@ -77,7 +282,7 @@ mod tests {
         let result = instance.get_or_prompt_project_name();
 
         assert!(result.is_ok());
-        assert_eq!(result.unwrap(), "my-graphql-api");
+        assert_eq!(result.unwrap(), "My API");
     }
 
     // Default trait implementation tests
